Add files via upload
In this paper, we have worked on finding the probability of a team/player winning in a tournament, conducted in a single-elimination Knockout format. The paper adopts a sequential approach of solving the problem, where it tries to find out the solution initially without any time constraints (brute force approach), and later optimizes the time taken.